Remove unnecessary threading.
There was two threads that executed sequentially and handoff between
them. This commit simplifies that to a single thread that gets right
down to business.
